Crate js

Modules

conversions
Conversions of Rust values to and from JSVal.

error
Functions to throw JavaScript exceptions from Rust.

glue

jsapi

jsval

panic

rust
Rust wrappers around the raw JS apis

typedarray
High-level, safe bindings for JS typed array APIs. Allows creating new typed arrays or wrapping existing JS reflectors, and prevents reinterpreting existing buffers as different types except in well-defined cases.

Macros

new_jsjitinfo_bitfield_1
Like JSJitInfo::new_bitfield_1, but usable in const contexts.

rooted

typedarray

Constants

JSCLASS_GLOBAL_SLOT_COUNT

JSCLASS_HIGH_FLAGS_SHIFT

JSCLASS_IS_DOMJSCLASS

JSCLASS_IS_GLOBAL

JSCLASS_IS_PROXY

JSCLASS_RESERVED_SLOTS_MASK

JSCLASS_USERBIT1

default_heapsize

Functions

JS_ARGV⚠

JS_CALLEE⚠