js::jsapi::BuildStackString

pub unsafe extern fn BuildStackString(cx: *mut JSContext, stack: HandleObject, stringp: MutableHandleString, indent: usize) -> bool

Given a SavedFrame JSObject stack, stringify it in the same format as Error.prototype.stack. The stringified stack out parameter is placed in the cx's compartment. Defaults to the empty string.

The same notes above about SavedFrame accessors applies here as well: cx doesn't need to be in stack's compartment, and stack can be null, a SavedFrame object, or a wrapper (CCW or Xray) around a SavedFrame object.

Optional indent parameter specifies the number of white spaces to indent each line.