Function js::jsapi::CompileFunction

pub unsafe extern "C" fn CompileFunction(
    cx: *mut JSContext, 
    scopeChain: *mut AutoObjectVector, 
    options: *const ReadOnlyCompileOptions, 
    name: *const c_char, 
    nargs: c_uint, 
    argnames: *const *const c_char, 
    chars: *const c_ushort, 
    length: usize, 
    fun: MutableHandleFunction
) -> bool

Compile a function with scopeChain plus the global as its scope chain. scopeChain must contain objects in the current compartment of cx. The actual scope chain used for the function will consist of With wrappers for those objects, followed by the current global of the compartment cx is in. This global must not be explicitly included in the scope chain.