js::jsapi::Evaluate3

pub unsafe extern fn Evaluate3(cx: *mut JSContext, scopeChain: *mut AutoObjectVector, options: *const ReadOnlyCompileOptions, chars: *const c_ushort, length: usize, rval: MutableHandleValue) -> bool

As above, but providing an explicit scope chain. scopeChain must not include the global object on it; that's implicit. It needs to contain the other objects that should end up on the script's scope chain.