js::jsapi::JS_GetFunctionDisplayId

pub unsafe extern fn JS_GetFunctionDisplayId(fun: *mut JSFunction) -> *mut JSString

Return a function's display name. This is the defined name if one was given where the function was defined, or it could be an inferred name by the JS engine in the case that the function was defined to be anonymous. This can still return nullptr if a useful display name could not be inferred. The same restrictions on rooting as those in JS_GetFunctionId apply.