Function js::jsapi::JS_GetPropertyDescriptorById

pub unsafe extern "C" fn JS_GetPropertyDescriptorById(
    cx: *mut JSContext, 
    obj: HandleObject, 
    id: HandleId, 
    desc: MutableHandle<PropertyDescriptor>
) -> bool

Like JS_GetOwnPropertyDescriptorById, but also searches the prototype chain if no own property is found directly on obj. The object on which the property is found is returned in desc.object(). If the property is not found on the prototype chain, this returns true with desc.object() set to null.