Struct js::jsapi::CompartmentCreationOptions

pub struct CompartmentCreationOptions {
    pub addonId_: *mut JSAddonId,
    pub traceGlobal_: JSTraceOp,
    pub zone_: CompartmentCreationOptions_jsapi_h_unnamed_8,
    pub invisibleToDebugger_: bool,
    pub mergeable_: bool,
    pub preserveJitCode_: bool,
    pub cloneSingletons_: bool,
    pub experimentalDateTimeFormatFormatToPartsEnabled_: bool,
    pub sharedMemoryAndAtomics_: bool,
    pub secureContext_: bool,
}

CompartmentCreationOptions specifies options relevant to creating a new compartment, that are either immutable characteristics of that compartment or that are discarded after the compartment has been created.

Access to these options on an existing compartment is read-only: if you need particular selections, make them before you create the compartment.

Fields
addonId_: *mut JSAddonId traceGlobal_: JSTraceOp zone_: CompartmentCreationOptions_jsapi_h_unnamed_8 invisibleToDebugger_: bool mergeable_: bool preserveJitCode_: bool cloneSingletons_: bool experimentalDateTimeFormatFormatToPartsEnabled_: bool sharedMemoryAndAtomics_: bool secureContext_: bool
Methods

impl CompartmentCreationOptions

unsafe fn setZone(&mut self,
spec: ZoneSpecifier)
-> *mut CompartmentCreationOptions

unsafe fn setSameZoneAs(&mut self,
obj: *mut JSObject)
-> *mut CompartmentCreationOptions

unsafe fn getSharedMemoryAndAtomicsEnabled(&mut self) -> bool

unsafe fn setSharedMemoryAndAtomicsEnabled(&mut self,
flag: bool)
-> *mut CompartmentCreationOptions