Struct js::jsapi::JSAutoStructuredCloneBuffer

pub struct JSAutoStructuredCloneBuffer {
    pub data_: *mut u64,
    pub nbytes_: usize,
    pub version_: u32,
    pub ownTransferables_: JSAutoStructuredCloneBuffer_StructuredClone_h_unnamed_8,
    pub callbacks_: *const JSStructuredCloneCallbacks,
    pub closure_: *mut c_void,
}

RAII sugar for JS_WriteStructuredClone.

Fields
data_: *mut u64 nbytes_: usize version_: u32 ownTransferables_: JSAutoStructuredCloneBuffer_StructuredClone_h_unnamed_8 callbacks_: *const JSStructuredCloneCallbacks closure_: *mut c_void
Methods

impl JSAutoStructuredCloneBuffer

unsafe fn data(&mut self) -> *mut u64

unsafe fn nbytes(&mut self) -> usize

unsafe fn clear(&mut self, optionalCallbacks: *const JSStructuredCloneCallbacks, closure: *mut c_void)

unsafe fn copy(&mut self, data: *const u64, nbytes: usize, version: u32, callbacks: *const JSStructuredCloneCallbacks, closure: *mut c_void) -> bool

Copy some memory. It will be automatically freed by the destructor.

unsafe fn adopt(&mut self, data: *mut u64, nbytes: usize, version: u32, callbacks: *const JSStructuredCloneCallbacks, closure: *mut c_void)

Adopt some memory. It will be automatically freed by the destructor. data must have been allocated by the JS engine (e.g., extracted via JSAutoStructuredCloneBuffer::steal).

unsafe fn steal(&mut self, datap: *mut *mut u64, nbytesp: *mut usize, versionp: *mut u32, callbacks: *mut *const JSStructuredCloneCallbacks, closure: *mut *mut c_void)

Release the buffer and transfer ownership to the caller. The caller is responsible for calling JS_ClearStructuredClone or feeding the memory back to JSAutoStructuredCloneBuffer::adopt.

unsafe fn abandon(&mut self)

Abandon ownership of any transferable objects stored in the buffer, without freeing the buffer itself. Useful when copying the data out into an external container, though note that you will need to use adopt() or JS_ClearStructuredClone to properly release that data eventually.

unsafe fn read(&mut self, cx: *mut JSContext, vp: MutableHandleValue, optionalCallbacks: *const JSStructuredCloneCallbacks, closure: *mut c_void) -> bool

unsafe fn write(&mut self, cx: *mut JSContext, v: HandleValue, optionalCallbacks: *const JSStructuredCloneCallbacks, closure: *mut c_void) -> bool

unsafe fn write1(&mut self, cx: *mut JSContext, v: HandleValue, transferable: HandleValue, optionalCallbacks: *const JSStructuredCloneCallbacks, closure: *mut c_void) -> bool