Struct js::jsapi::JSFunctionSpec

pub struct JSFunctionSpec {
    pub name: *const c_char,
    pub call: JSNativeWrapper,
    pub nargs: u16,
    pub flags: u16,
    pub selfHostedName: *const c_char,
}

To define a native function, set call to a JSNativeWrapper. To define a self-hosted function, set selfHostedName to the name of a function compiled during JSRuntime::initSelfHosting.

Fields

name

call

nargs

flags

selfHostedName