Struct js::jsapi::JSJitInfo

#[repr(C)]

pub struct JSJitInfo {
    pub call: *const c_void,
    pub protoID: u16,
    pub depth: u16,
    pub _bitfield_1: u32,
}

This struct contains metadata passed from the DOM to the JS Engine for JIT optimizations on DOM property accessors. Eventually, this should be made available to general JSAPI users, but we are not currently ready to do so.

Fields
call: *const c_void protoID: u16 depth: u16 _bitfield_1: u32

The OpType that says what sort of function we are.

Methods

impl JSJitInfo

pub fn type_(&self) -> u32

pub fn set_type_(&mut self, val: u8)

pub fn aliasSet_(&self) -> u32

pub fn set_aliasSet_(&mut self, val: u8)

pub fn returnType_(&self) -> u32

pub fn set_returnType_(&mut self, val: u8)

pub fn isInfallible(&self) -> u32

pub fn set_isInfallible(&mut self, val: bool)

pub fn isMovable(&self) -> u32

pub fn set_isMovable(&mut self, val: bool)

pub fn isEliminatable(&self) -> u32

pub fn set_isEliminatable(&mut self, val: bool)

pub fn isAlwaysInSlot(&self) -> u32

pub fn set_isAlwaysInSlot(&mut self, val: bool)

pub fn isLazilyCachedInSlot(&self) -> u32

pub fn set_isLazilyCachedInSlot(&mut self, val: bool)

pub fn isTypedMethod(&self) -> u32

pub fn set_isTypedMethod(&mut self, val: bool)

pub fn slotIndex(&self) -> u32

pub fn set_slotIndex(&mut self, val: u16)

pub fn new_bitfield_1(
type_: u8,
aliasSet_: u8,
returnType_: u8,
isInfallible: bool,
isMovable: bool,
isEliminatable: bool,
isAlwaysInSlot: bool,
isLazilyCachedInSlot: bool,
isTypedMethod: bool,
slotIndex: u16
) -> u32