Struct js::jsapi::MutableHandle

pub struct MutableHandle<T> {
    pub _base: MutableHandleBase<T>,
    pub ptr: *mut T,
}

Similar to a handle, but the underlying storage can be changed. This is useful for outparams.

If you want to add additional methods to MutableHandle for a specific specialization, define a MutableHandleBase specialization containing them.

Fields

_base

ptr

Methods

impl<T: Copy> MutableHandle<T>

unsafe fn from_marked_location(ptr: *mut T) -> MutableHandle<T>

fn to_handle(&self) -> Handle<T>

impl<T: Copy> MutableHandle<T>

fn get(&self) -> T

fn set(&self, v: T)

fn handle(&self) -> Handle<T>