Struct js::jsapi::Object

pub struct Object {
    pub group: *mut ObjectGroup,
    pub shape: *mut Shape,
    pub slots: *mut Value,
    pub _1: *mut c_void,
}

This layout is shared by all native objects. For non-native objects, the group may always be accessed safely, and other members may be as well, depending on the object's specific layout.

Fields

group

shape

slots

_1

Methods

impl Object

unsafe fn numFixedSlots(&mut self) -> usize

unsafe fn fixedSlots(&mut self) -> *mut Value

unsafe fn slotRef(&mut self, slot: usize) -> *mut Value