Struct js::jsapi::PerformanceGroup

#[repr(C)]

pub struct PerformanceGroup {
    pub _vftable: *const _vftable_PerformanceGroup,
    pub recentCycles_: u64,
    pub recentTicks_: u64,
    pub recentCPOW_: u64,
    pub iteration_: u64,
    pub isActive_: bool,
    pub isUsedInThisIteration_: bool,
    pub owner_: *const AutoStopwatch,
    pub refCount_: u64,
}

Abstract base class for a representation of the performance of a component. Embeddings interested in performance monitoring should provide a concrete implementation of this class, as well as the relevant callbacks (see below).

Fields
_vftable: *const _vftable_PerformanceGroup recentCycles_: u64 recentTicks_: u64 recentCPOW_: u64 iteration_: u64 isActive_: bool isUsedInThisIteration_: bool owner_: *const AutoStopwatch refCount_: u64
Methods

impl PerformanceGroup

pub unsafe fn iteration(&mut self) -> u64

pub unsafe fn isAcquired(&mut self, it: u64) -> bool

pub unsafe fn isAcquired1(
&mut self,
it: u64,
owner: *const AutoStopwatch
) -> bool

pub unsafe fn acquire(&mut self, it: u64, owner: *const AutoStopwatch)

pub unsafe fn release(&mut self, it: u64, owner: *const AutoStopwatch)

pub unsafe fn recentCycles(&mut self, iteration: u64) -> u64

pub unsafe fn addRecentCycles(&mut self, iteration: u64, cycles: u64)

pub unsafe fn recentTicks(&mut self, iteration: u64) -> u64

pub unsafe fn addRecentTicks(&mut self, iteration: u64, ticks: u64)

pub unsafe fn recentCPOW(&mut self, iteration: u64) -> u64

pub unsafe fn addRecentCPOW(&mut self, iteration: u64, CPOW: u64)

pub unsafe fn resetRecentData(&mut self)

pub unsafe fn isActive(&mut self) -> bool

pub unsafe fn setIsActive(&mut self, arg1: bool)

pub unsafe fn isUsedInThisIteration(&mut self) -> bool

pub unsafe fn setIsUsedInThisIteration(&mut self, arg1: bool)

pub unsafe fn AddRef(&mut self)

pub unsafe fn Release(&mut self)