Struct js::jsapi::Rooted

pub struct Rooted<T> {
    pub _base: RootedBase<T>,
    pub stack: *mut *mut Rooted<*mut c_void>,
    pub prev: *mut Rooted<*mut c_void>,
    pub ptr: T,
}

Local variable of type T whose value is always rooted. This is typically used for local variables, or for non-rooted values being passed to a function that requires a handle, e.g. Foo(Root(cx, x)).

If you want to add additional methods to Rooted for a specific specialization, define a RootedBase specialization containing them.

Fields
_base: RootedBase<T> stack: *mut *mut Rooted<*mut c_void> prev: *mut Rooted<*mut c_void> ptr: T
Methods

impl<T: RootKind + Copy> Rooted<T>

fn new_with_addr(cx: *mut JSContext, initial: T, addr: *const u8) -> Rooted<T>

fn new(cx: *mut JSContext, initial: T) -> Rooted<T>

fn handle(&self) -> Handle<T>

fn handle_mut(&mut self) -> MutableHandle<T>