Struct js::jsapi::Value

pub struct Value {
    pub data: jsval_layout,
}

JS::Value is the interface for a single JavaScript Engine value. A few general notes on JS::Value:

JS::Value has setX() and isX() members for X in

{ Int32, Double, String, Symbol, Boolean, Undefined, Null, Object, Magic }

JS::Value also contains toX() for each of the non-singleton types.

Magic is a singleton type whose payload contains either a JSWhyMagic "reason" for the magic value or a uint32_t value. By providing JSWhyMagic values when creating and checking for magic values, it is possible to assert, at runtime, that only magic values with the expected reason flow through a particular value. For example, if cx->exception has a magic value, the reason must be JS_GENERATOR_CLOSING.

The JS::Value operations are preferred. The JSVAL_* operations remain for compatibility; they may be removed at some point. These operations mostly provide similar functionality. But there are a few key differences. One is that JS::Value gives null a separate type. Also, to help prevent mistakenly boxing a nullable JSObject* as an object, Value::setObject takes a JSObject&. (Conversely, Value::toObject returns a JSObject&.) A convenience member Value::setObjectOrNull is provided.

JSVAL_VOID is the same as the singleton value of the Undefined type.

Note that JS::Value is 8 bytes on 32 and 64-bit architectures. Thus, on 32-bit user code should avoid copying jsval/JS::Value as much as possible, preferring to pass by const Value&.

Fields

data

Methods

impl Value

unsafe fn setNull(&mut self)

unsafe fn setUndefined(&mut self)

unsafe fn setInt32(&mut self, i: i32)

unsafe fn getInt32Ref(&mut self) -> *mut i32

unsafe fn setDouble(&mut self, d: f64)

unsafe fn setNaN(&mut self)

unsafe fn getDoubleRef(&mut self) -> *mut f64

unsafe fn setString(&mut self, str: *mut JSString)

unsafe fn setSymbol(&mut self, sym: *mut Symbol)

unsafe fn setObject(&mut self, obj: *mut JSObject)

unsafe fn setBoolean(&mut self, b: bool)

unsafe fn setMagic(&mut self, why: JSWhyMagic)

unsafe fn setMagicUint32(&mut self, payload: u32)

unsafe fn setNumber(&mut self, ui: u32) -> bool

unsafe fn setNumber1(&mut self, d: f64) -> bool

unsafe fn setObjectOrNull(&mut self, arg: *mut JSObject)

unsafe fn swap(&mut self, rhs: *mut Value)

unsafe fn isUndefined(&mut self) -> bool

unsafe fn isNull(&mut self) -> bool

unsafe fn isNullOrUndefined(&mut self) -> bool

unsafe fn isInt32(&mut self) -> bool

unsafe fn isInt321(&mut self, i32: i32) -> bool

unsafe fn isDouble(&mut self) -> bool

unsafe fn isNumber(&mut self) -> bool

unsafe fn isString(&mut self) -> bool

unsafe fn isSymbol(&mut self) -> bool

unsafe fn isObject(&mut self) -> bool

unsafe fn isPrimitive(&mut self) -> bool

unsafe fn isObjectOrNull(&mut self) -> bool

unsafe fn isGCThing(&mut self) -> bool

unsafe fn isBoolean(&mut self) -> bool

unsafe fn isTrue(&mut self) -> bool

unsafe fn isFalse(&mut self) -> bool

unsafe fn isMagic(&mut self) -> bool

unsafe fn isMagic1(&mut self, why: JSWhyMagic) -> bool

unsafe fn isMarkable(&mut self) -> bool

unsafe fn traceKind(&mut self) -> TraceKind

unsafe fn whyMagic(&mut self) -> JSWhyMagic

unsafe fn magicUint32(&mut self) -> u32

unsafe fn toInt32(&mut self) -> i32

unsafe fn toDouble(&mut self) -> f64

unsafe fn toNumber(&mut self) -> f64

unsafe fn toString(&mut self) -> *mut JSString

unsafe fn toSymbol(&mut self) -> *mut Symbol

unsafe fn toObject(&mut self) -> *mut JSObject

unsafe fn toObjectOrNull(&mut self) -> *mut JSObject

unsafe fn toGCThing(&mut self) -> *mut Cell

unsafe fn toGCCellPtr(&mut self) -> GCCellPtr

unsafe fn toBoolean(&mut self) -> bool

unsafe fn payloadAsRawUint32(&mut self) -> u32

unsafe fn asRawBits(&mut self) -> u64

unsafe fn extractNonDoubleType(&mut self) -> JSValueType

unsafe fn setPrivate(&mut self, ptr: *mut c_void)

unsafe fn toPrivate(&mut self) -> *mut c_void

unsafe fn setPrivateUint32(&mut self, ui: u32)

unsafe fn toPrivateUint32(&mut self) -> u32

unsafe fn setUnmarkedPtr(&mut self, ptr: *mut c_void)

unsafe fn toUnmarkedPtr(&mut self) -> *mut c_void

unsafe fn payloadWord(&mut self) -> *const usize

unsafe fn payloadUIntPtr(&mut self) -> *const usize