Type Definition js::jsapi::FreeTransferStructuredCloneOp

type FreeTransferStructuredCloneOp = Option<unsafe extern "C" fn(tag: u32, ownership: TransferableOwnership, content: *mut c_void, extraData: u64, closure: *mut c_void)>;

Called when JS_ClearStructuredClone has to free an unknown transferable object. Note that it should never trigger a garbage collection (and will assert in a debug build if it does.)