Type Definition js::jsapi::JSDeletePropertyOp

type JSDeletePropertyOp = Option<unsafe extern "C" fn(cx: *mut JSContext, obj: HandleObject, id: HandleId, result: *mut ObjectOpResult) -> bool>;

Delete a property named by id in obj.

If an error occurred, return false as per normal JSAPI error practice.

If no error occurred, but the deletion attempt wasn't allowed (perhaps because the property was non-configurable), call result.fail() and return true. This will cause |delete obj[id]| to evaluate to false in non-strict mode code, and to throw a TypeError in strict mode code.

If no error occurred and the deletion wasn't disallowed (this is not the same as saying that a deletion actually occurred -- deleting a non-existent property, or an inherited property, is allowed -- it's just pointless), call result.succeed() and return true.