Module js::typedarray

High-level, safe bindings for JS typed array APIs. Allows creating new typed arrays or wrapping existing JS reflectors, and prevents reinterpreting existing buffers as different types except in well-defined cases.

Structs

ArrayBufferU8
A kind of typed array.

ArrayBufferViewU8
A kind of typed array.

ClampedU8
A kind of typed array.

Float32
A kind of typed array.

Float64
A kind of typed array.

Int8
A kind of typed array.

Int16
A kind of typed array.

Int32
A kind of typed array.

TypedArray
A rooted typed array.

Uint8
A kind of typed array.

Uint16
A kind of typed array.

Uint32
A kind of typed array.

Enums

CreateWith

Traits

TypedArrayElement
Internal trait used to associate an element type with an underlying representation and various functions required to manipulate typed arrays of that element type.

TypedArrayElementCreator
Internal trait for creating new typed arrays.

Type Definitions

ArrayBuffer
The ArrayBuffer type.

ArrayBufferView
The ArrayBufferView type

Float32Array
The Float32Array type.

Float64Array
The Float64Array type.

Int16Array
The Int16Array type.

Int32Array
The Int32Array type.

Int8Array
The Int8Array type.

Uint16Array
The Uint16Array type.

Uint32Array
The Uint32Array type.

Uint8Array
The Uint8Array type.

Uint8ClampedArray
The Uint8ClampedArray type.