Struct js::typedarray::TypedArray

pub struct TypedArray<'a, T: 'a + TypedArrayElement> {
    object: RootedGuard<'a, *mut JSObject>,
    computed: Option<(*mut T::Element, u32)>,
}

A rooted typed array.

Fields
object: RootedGuard<'a, *mut JSObject> computed: Option<(*mut T::Element, u32)>
Methods

impl<'a, T: TypedArrayElement> TypedArray<'a, T>

fn from(cx: *mut JSContext,
root: &'a mut Rooted<*mut JSObject>,
object: *mut JSObject)
-> Result<Self, ()>

Create a typed array representation that wraps an existing JS reflector. This operation will fail if attempted on a JS object that does not match the expected typed array details.

fn data(&mut self) -> (*mut T::Element, u32)

unsafe fn as_slice(&mut self) -> &[T::Element]

Unsafety

The returned slice can be invalidated if the underlying typed array is neutered.

unsafe fn as_mut_slice(&mut self) -> &mut [T::Element]

Unsafety

The returned slice can be invalidated if the underlying typed array is neutered.

The underlying JSObject can be aliased, which can lead to Undefined Behavior due to mutable aliasing.

impl<'a, T: TypedArrayElementCreator + TypedArrayElement> TypedArray<'a, T>

unsafe fn create(cx: *mut JSContext,
with: CreateWith<T::Element>,
result: MutableHandleObject)
-> Result<(), ()>

Create a new JS typed array, optionally providing initial data that will be copied into the newly-allocated buffer. Returns the new JS reflector.

unsafe fn update(&mut self, data: &[T::Element])

Update an existed JS typed array

unsafe fn update_raw(data: &[T::Element], result: HandleObject)