Struct js::jsapi::AutoStableStringChars

pub struct AutoStableStringChars {
    pub s_: RootedString,
    pub AutoStableStringChars_jsfriendapi_h_unnamed_14: AutoStableStringChars_jsfriendapi_h_unnamed_14,
    pub state_: AutoStableStringChars_State,
    pub ownsChars_: bool,
}

This class provides safe access to a string's chars across a GC. Once we allocate strings and chars in the nursery (bug 903519), this class will have to make a copy of the string's chars if they are allocated in the nursery, so it's best to avoid using this class unless you really need it. It's usually more efficient to use the latin1Chars/twoByteChars JSString methods and often the code can be rewritten so that only indexes instead of char pointers are used in parts of the code that can GC.

Fields
s_: RootedString AutoStableStringChars_jsfriendapi_h_unnamed_14: AutoStableStringChars_jsfriendapi_h_unnamed_14 state_: AutoStableStringChars_State ownsChars_: bool
Methods

impl AutoStableStringChars

unsafe fn init(&mut self, cx: *mut JSContext, s: *mut JSString) -> bool

unsafe fn initTwoByte(&mut self, cx: *mut JSContext, s: *mut JSString) -> bool