Struct js::jsapi::CompartmentCreationOptions

pub struct CompartmentCreationOptions {
    pub addonId_: *mut JSAddonId,
    pub traceGlobal_: JSTraceOp,
    pub zone_: CompartmentCreationOptions_jsapi_h_unnamed_9,
    pub invisibleToDebugger_: bool,
    pub mergeable_: bool,
    pub preserveJitCode_: bool,
    pub cloneSingletons_: bool,
    pub experimentalDateTimeFormatFormatToPartsEnabled_: bool,
    pub sharedMemoryAndAtomics_: bool,
}

CompartmentCreationOptions specifies options relevant to creating a new compartment, that are either immutable characteristics of that compartment or that are discarded after the compartment has been created.

Access to these options on an existing compartment is read-only: if you need particular selections, make them before you create the compartment.

Fields

addonId_

traceGlobal_

zone_

invisibleToDebugger_

mergeable_

preserveJitCode_

cloneSingletons_

experimentalDateTimeFormatFormatToPartsEnabled_

sharedMemoryAndAtomics_

Methods

impl CompartmentCreationOptions

unsafe fn addonIdOrNull(&mut self) -> *mut JSAddonId

unsafe fn setAddonId(&mut self, id: *mut JSAddonId) -> *mut CompartmentCreationOptions

unsafe fn getTrace(&mut self) -> JSTraceOp

unsafe fn setTrace(&mut self, op: JSTraceOp) -> *mut CompartmentCreationOptions

unsafe fn zonePointer(&mut self) -> *mut c_void

unsafe fn zoneSpecifier(&mut self) -> ZoneSpecifier

unsafe fn setZone(&mut self, spec: ZoneSpecifier) -> *mut CompartmentCreationOptions

unsafe fn setSameZoneAs(&mut self, obj: *mut JSObject) -> *mut CompartmentCreationOptions

unsafe fn invisibleToDebugger(&mut self) -> bool

unsafe fn setInvisibleToDebugger(&mut self, flag: bool) -> *mut CompartmentCreationOptions

unsafe fn mergeable(&mut self) -> bool

unsafe fn setMergeable(&mut self, flag: bool) -> *mut CompartmentCreationOptions

unsafe fn preserveJitCode(&mut self) -> bool

unsafe fn setPreserveJitCode(&mut self, flag: bool) -> *mut CompartmentCreationOptions

unsafe fn cloneSingletons(&mut self) -> bool

unsafe fn setCloneSingletons(&mut self, flag: bool) -> *mut CompartmentCreationOptions

unsafe fn experimentalDateTimeFormatFormatToPartsEnabled(&mut self) -> bool

unsafe fn setExperimentalDateTimeFormatFormatToPartsEnabled(&mut self, flag: bool) -> *mut CompartmentCreationOptions

unsafe fn getSharedMemoryAndAtomicsEnabled(&mut self) -> bool

unsafe fn setSharedMemoryAndAtomicsEnabled(&mut self, flag: bool) -> *mut CompartmentCreationOptions