Struct js::jsapi::JSJitInfo

pub struct JSJitInfo {
    pub call: *const c_void,
    pub protoID: u16,
    pub depth: u16,
    pub _bitfield_1: u32,
}

This struct contains metadata passed from the DOM to the JS Engine for JIT optimizations on DOM property accessors. Eventually, this should be made available to general JSAPI users, but we are not currently ready to do so.

Fields

call

protoID

depth

_bitfield_1

The OpType that says what sort of function we are.

Methods

impl JSJitInfo

fn set_type_(&mut self, val: u8)

fn set_aliasSet_(&mut self, val: u8)

fn set_returnType_(&mut self, val: u8)

fn set_isInfallible(&mut self, val: bool)

fn set_isMovable(&mut self, val: bool)

fn set_isEliminatable(&mut self, val: bool)

fn set_isAlwaysInSlot(&mut self, val: bool)

fn set_isLazilyCachedInSlot(&mut self, val: bool)

fn set_isTypedMethod(&mut self, val: bool)

fn set_slotIndex(&mut self, val: u16)

const fn new_bitfield_1(type_: u8, aliasSet_: u8, returnType_: u8, isInfallible: bool, isMovable: bool, isEliminatable: bool, isAlwaysInSlot: bool, isLazilyCachedInSlot: bool, isTypedMethod: bool, slotIndex: u16) -> u32

impl JSJitInfo

unsafe fn needsOuterizedThisObject(&mut self) -> bool

unsafe fn isTypedMethodJitInfo(&mut self) -> bool

unsafe fn type_(&mut self) -> JSJitInfo_OpType

unsafe fn aliasSet(&mut self) -> JSJitInfo_AliasSet

unsafe fn returnType(&mut self) -> JSValueType