Struct js::jsapi::PerformanceGroup

pub struct PerformanceGroup {
    pub _vftable: *const _vftable_PerformanceGroup,
    pub recentCycles_: u64,
    pub recentTicks_: u64,
    pub recentCPOW_: u64,
    pub iteration_: u64,
    pub isActive_: bool,
    pub isUsedInThisIteration_: bool,
    pub owner_: *const AutoStopwatch,
    pub refCount_: u64,
}

Abstract base class for a representation of the performance of a component. Embeddings interested in performance monitoring should provide a concrete implementation of this class, as well as the relevant callbacks (see below).

Fields

_vftable

recentCycles_

recentTicks_

recentCPOW_

iteration_

isActive_

isUsedInThisIteration_

owner_

refCount_

Methods

impl PerformanceGroup

unsafe fn iteration(&mut self) -> u64

unsafe fn isAcquired(&mut self, it: u64) -> bool

unsafe fn isAcquired1(&mut self, it: u64, owner: *const AutoStopwatch) -> bool

unsafe fn acquire(&mut self, it: u64, owner: *const AutoStopwatch)

unsafe fn release(&mut self, it: u64, owner: *const AutoStopwatch)

unsafe fn recentCycles(&mut self, iteration: u64) -> u64

unsafe fn addRecentCycles(&mut self, iteration: u64, cycles: u64)

unsafe fn recentTicks(&mut self, iteration: u64) -> u64

unsafe fn addRecentTicks(&mut self, iteration: u64, ticks: u64)

unsafe fn recentCPOW(&mut self, iteration: u64) -> u64

unsafe fn addRecentCPOW(&mut self, iteration: u64, CPOW: u64)

unsafe fn resetRecentData(&mut self)

unsafe fn isActive(&mut self) -> bool

unsafe fn setIsActive(&mut self, arg1: bool)

unsafe fn isUsedInThisIteration(&mut self) -> bool

unsafe fn setIsUsedInThisIteration(&mut self, arg1: bool)

unsafe fn AddRef(&mut self)

unsafe fn Release(&mut self)