Module js::rust

Rust wrappers around the raw JS apis

Structs

AutoObjectVectorWrapper

CompileOptionsWrapper

Runtime
A wrapper for the JSRuntime and JSContext structures in SpiderMonkey.

Constants

ChunkLocationBitNursery

ChunkLocationOffset

ChunkMask

ChunkShift

ChunkSize

ConstNullValue

STACK_QUOTA

SYSTEM_CODE_BUFFER

TRUSTED_SCRIPT_BUFFER

Statics

SIMPLE_GLOBAL_CLASS
This is a simple JSClass for global objects, primarily intended for tests.

Traits

GCMethods

RootKind

ToResult

Functions

IsInsideNursery

ToBoolean

ToInt32

ToInt64

ToNumber

ToString

ToUint16

ToUint32

ToUint64

convert_from_int32

define_methods
Defines methods on obj. The last entry of methods must contain zeroed memory.

define_properties
Defines attributes on obj. The last entry of properties must contain zeroed memory.

reportError